smart360_API/requirements.txt
Django==3.0.3
PyJWT==1.7.1
djangorestframework==3.11.0
django-cors-headers==3.2.1